+++ ruoyi-modules/ruoyi-zlm/src/main/java/com/ruoyi/zlm/service/IInviteStreamService.java
@@ -23,12 +23,12 @@ public interface IInviteStreamService {
     /**
      * 获取点播的状态信息
      */
-    InviteInfo getInviteInfo(InviteSessionType type, Integer channelId, String stream);
+    InviteInfo getInviteInfo(InviteSessionType type, Long channelId, String stream);
 
     /**
      * 移除点播的状态信息
      */
-    void removeInviteInfo(InviteSessionType type, Integer channelId, String stream);
+    void removeInviteInfo(InviteSessionType type, Long channelId, String stream);
 
     /**
      * 移除点播的状态信息
@@ -38,14 +38,14 @@ public interface IInviteStreamService {
     /**
      * 移除点播的状态信息
      */
-    void removeInviteInfoByDeviceAndChannel(InviteSessionType inviteSessionType, Integer channelId);
+    void removeInviteInfoByDeviceAndChannel(InviteSessionType inviteSessionType, Long channelId);
 
     List<InviteInfo> getAllInviteInfo();
 
     /**
      * 获取点播的状态信息
      */
-    InviteInfo getInviteInfoByDeviceAndChannel(InviteSessionType type, Integer channelId);
+    InviteInfo getInviteInfoByDeviceAndChannel(InviteSessionType type, Long channelId);
 
     /**
      * 获取点播的状态信息
@@ -56,12 +56,12 @@ public interface IInviteStreamService {
     /**
      * 添加一个invite回调
      */
-    void once(InviteSessionType type, Integer channelId, String stream, ErrorCallback<StreamInfo> callback);
+    void once(InviteSessionType type, Long channelId, String stream, ErrorCallback<StreamInfo> callback);
 
     /**
      * 调用一个invite回调
      */
-    void call(InviteSessionType type, Integer channelId, String stream, int code, String msg, StreamInfo data);
+    void call(InviteSessionType type, Long channelId, String stream, int code, String msg, StreamInfo data);
 
     /**
      * 清空一个设备的所有invite信息
